Add files via upload
Nigeria's Presidential Elections was conducted on the 25th of february 2023. Citizens across the country went out to cast their votes for their preferred candidates. PDP's Atiku Abubakar, LP's Peter Obi and APC's Bola Ahmed tinubu are the 3 candidated projected to be at the forefront of the race. However, there is an argument, 'To what extent does Social media determine the outcome of an election ?'. All tweets with relevant hashtags related to each candidate was extracted from twitter on that election day (25/02/2023). Twitter #hashtag has decided !. Will there be a correlation with that of INEC, the electoral body?.